+++ daemon/internal/handlers/system_extended.go
@@ -194,7 +194,7 @@ func (h *SnapshotScheduleHandler) regenerateCron(schedules []SnapshotSchedule) {
 		// we escape any single quotes in the json payload (though there shouldn't be any now).
 		safePayload := strings.ReplaceAll(payload, "'", "'\\''")
 		mainCmd := fmt.Sprintf(
-			`curl -sf -X POST http://127.0.0.1:9000/api/zfs/snapshots/cron-hook -H 'Content-Type: application/json' -H 'X-Internal-Token: %s' -d '%s'`,
+			`curl -sf --unix-socket /run/dplaneos/dplaned.sock -X POST http://localhost/api/zfs/snapshots/cron-hook -H 'Content-Type: application/json' -H 'X-Internal-Token: %s' -d '%s'`,
 			cronToken,
 			safePayload,
 		)